oracle/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
substrate-interface>=1.4.0
aiohttp==3.7.4
jsonpath-ng==1.5.3
pydantic==1.8.2
python-multipart==0.0.5
web3==6.11.1
pytest==6.2.5
pytest-asyncio==0.15.1
prometheus-client==0.11.0
asyncio>=3.4.3
typing-extensions>=4.5.0
pytest-cov==2.12.1
mypy>=1.0.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
numpy==1.21.2
requests==2.26.0 